requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
nltk==3.8.1
pandas==2.0.3